+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [x.y.z] (Unreleased)
 
+## Added
+
+- Prevent enemies from knowing where sneaking chars are during the ambush, unless they pass their perception check against the char's stealth 
+
 ## Changed
 
 - Internal refactor to support pre- and post- ambush functions for more flexibility later